+++ seleniumpm/examples/superwikipedia.py
@@ -1,4 +1,5 @@
 from seleniumpm.examples.wikipedia import Wikipedia
+from seleniumpm.examples.widgets.WikipediaPersonal import WikipediaPersonal
 from seleniumpm.webelements.link import Link
 from seleniumpm.locator import Locator
 from selenium.webdriver.common.by import By
@@ -15,7 +16,4 @@ class SuperWikipedia(Wikipedia):
     def __init__(self, driver, url):
         super(SuperWikipedia, self).__init__(driver, url)
         self.mainpage_link = Link(driver, Locator(By.XPATH, "//li[@id='n-mainpage-description']/a"))
-        self.talk_link = Link(driver, Locator(By.XPATH, "//li[@id='pt-anontalk']/a"))
-        self.contributions_link = Link(driver, Locator(By.XPATH, "//li[@id='pt-anoncontribs']/a"))
-        self.createaccount_link = Link(driver, Locator(By.XPATH, "//li[@id='pt-createaccount']/a"))
-        self.login_link = Link(driver, Locator(By.XPATH, "//li[@id='pt-login']/a"))
+        self.personal_widget = WikipediaPersonal(driver, Locator(By.XPATH, "//div[@id='p-personal']"))