+++ seleniumpm/webelements/widget.py
@@ -1,4 +1,5 @@
 from seleniumpm.webelements.clickable import Clickable
+from seleniumpm.webelements.element import Element
 
 
 class Widget(Clickable):
@@ -7,3 +8,24 @@ class Widget(Clickable):
 
     def validate(self):
         raise NotImplementedError
+
+    def get_element_attr(self, type=Element):
+        """
+        Retrieves a list of WebElements on a Widget. Optionally, you can pass in a different type (e.g. Button,
+        Link, TextElement) to return only those types associated with a Widget object.
+        :param type: one of the seleniumpm.webelement types (Default: seleniumpm.webelements.Element)
+        :return: This is a list of attributes of base type seleniumpm.webelements.Element
+        """
+        elements = []
+        for attr in dir(self):
+            element = getattr(self, attr)
+            # Ensure that it is of type Element
+            if isinstance(element, Element):
+                # If it is a widget, then recursively drill down and get its Elements
+                if isinstance(element, Widget):
+                    for welement in element.get_element_attr(type=type):
+                        elements.append(welement)
+                # Add the element if it matches the expected type
+                if isinstance(element, type):
+                    elements.append(element)
+        return elements